Add gauss-seidel
add norms (not tested yet)
fix typo "double max_iter"
change all max_iters to size_t

+++ numerical/nonlinear-eqn/bisection.cpp
@@ -15,7 +15,7 @@ using namespace std;
  *
  * Ouput: An approximation to the root of f(x)=0 on [a,b].
  */
-double bisection(double f(double), double a, double b, double tol, double max_iter)
+double bisection(double f(double), double a, double b, double tol, size_t max_iter)
 {
   double c, ya = f(a), yb = f(b), yc;
   int i = 0;
@@ -57,7 +57,7 @@ int main()
   // Test
   double a = 0, b = 2;
   double tol = 1e-2;
-  int max_iter = 20;
+  size_t max_iter = 20;
 
   cout << "Solve ln(x+1) - cos(x) = 0 on [0,2] correct to within 1e-2 using the bisection method:" << endl;
   cout << "Solution: x = " << setiosflags(ios::fixed) << setprecision(6)
