+++ numerical/nonlinear-eqn/newtonraphson.cpp
@@ -16,7 +16,7 @@ using namespace std;
  *
  * Ouput: An approximation to the root of f(x)=0.
  */
-double newtonraphson(double f(double), double g(double), double z0, double tol, int max_iter)
+double newtonraphson(double f(double), double g(double), double z0, double tol, size_t max_iter)
 {
   double z, error;
   int i = 0;
@@ -51,7 +51,7 @@ int main()
   // Test
   double z0 = 1.5;
   double tol = 1e-4;
-  int max_iter = 20;
+  size_t max_iter = 20;
 
   cout << "Solve x^3 + 2x^2 - 5 = 0 on [1,2] correct to within 1e-4 using Newton-Raphson method:" << endl;
   cout << "Solution: x = " << setiosflags(ios::fixed) << setprecision(6)
